Renamed wcag-link-data-processor to wcag-link-data-indexer
