+++ src/wcag-link-data-indexer.ts
